requirements.txt
google-cloud-error-reporting==0.30.1
lxml
pyyaml
requests
pytest
flask

